university_admission_prediction/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-crispy-forms>=2.0
Pillow>=9.5.0
python-dotenv>=1.0.0
django-filter>=23.2
drf-yasg>=1.21.5
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.2
django-cors-headers>=4.1.0
google-cloud-aiplatform>=1.25.0
streamlit>=1.30.0
plotly>=5.18.0
pytest>=7.4.0
pytest-django>=4.5.2
coverage>=7.2.7
sphinx>=7.1.2
django-rest-swagger>=2.2.0
django-allauth>=0.54.0
tensorflow>=2.15.0
